Higor Cirne

+++ simulador-glosa/src/main/groovy/br/com/zgsolucoes/simuladorglosa/Dado.groovy
@@ -4,6 +4,6 @@ class Dado {
 
 	String codigo
 	String tipo
-	Long valor
+	BigDecimal valor
 
 }
